Laboratório III/requirements.txt
grpcio
grpcio-tools
protobuf
tk
tk-tools
Events
